Extract routes to separate files
